requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
python-multipart>=0.0.18
ifcopenshell>=0.8.1
pymupdf>=1.25.0
pytesseract>=0.3.13
fpdf2>=2.8.0
eval-type-backport>=0.2.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
